A library for generating a google API3 fields selector
Recent google API's have a "fields" attribute on requests that can be
populated to determine what parts of the response resource should be
populated. This allows clients request only the subset of information
they are interested in reducing bandwidth and overhead when receiving
the responses.

Most current libraries when interacting with the API therefore define
the response objects as a struct filled with Option's. Since the
user provides an arbitrary fields parameter the response that comes back
may have any number of the fields populated. Dealing with the response
is therefore tedious and the user needs to be aware of the fields that
were requested when inspecting the results. One response object may have
effectively different fields than another response object leading to
code that while technically type safe (they are Option's after all) you
don't get a lot of benefits that the type system could provide.

This library takes an alternative approach. The user of the library
defines what fields they want a response to contain by defining a
struct. The struct then implements a "FieldSelector" trait.
FieldSelector::field_selector() then returns a String that when provided
to the "fields" attribute of a google API request will request only the
fields defined in the struct.

This seems like a great way to conveniently define what fields the user
is interested in while bringing back a lot of the benefits of the type
system while also reducing the burden on the library author to
regenerate a new response object whenever google adds a field to a
resource.